refactor(ors_control_app): extract routes/agent
Move /api/agent/chat plus all of its supporting helpers into routes/agent.ts:
- ROUTING_SYSTEM_PROMPT, AGENT_PROFILE_ALIASES, AGENT_VALID_PROFILES
- TOOL_PROCEDURE_MAP, POI_CATEGORY_MAP
- normalizeAgentProfile, escAgentSql, escAgentSqlStr
- executeToolPoi, executeToolLocally
- AGENT_MODELS + agentModel state
- callCortexCompleteStreaming, callCortexComplete (with mistral fallback)
- findMatchingBrace, parseToolCall, callCortexAgentWithToolLoop
- sendSseEvent

server/index.ts mounts via app.use(createAgentRouter()).
Lines down 884 -> 525 (a 7.1x reduction from the original 3,726-line
monolith). Phase 1 of holistic-control-app-refactor-final, route extraction
step 8 (final route group). No behaviour change. tsc passes cleanly.
